seperated Infinite feature in a component
